pyproject.toml
[project]
name = "subtle"
version = "0.1.0"
description = "Subtitle generator from video."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "moviepy>=2.2.1",
    "openai-whisper>=20250625",
    "srt>=3.5.3",
    "streamlit>=1.51.0",
]
